api refactoring

+++ BookApp.API/Dtos/BookCatalogCreateDto.cs
@@ -1,7 +1,7 @@
 namespace BookApp.API.Dtos {
     public class BookCatalogCreateDto {
         public int BookId { get; set; }
-        public int? CatalogId { get; set; }
+        public long? CatalogId { get; set; }
         public string CatalogName { get; set; }
 
         public int UserId { get; set; }
